Warning fixing.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@30631 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/ptbl/xp/pp_Revision.cpp
@@ -284,7 +284,8 @@ std::string PP_Revision::toString() const
 
 bool PP_Revision::onlyContainsAbiwordChangeTrackingMarkup() const
 {
-    UT_DEBUGMSG(("onlyContainsAbiwordChangeTrackingMarkup(top) ac:%d pc:%d\n", getAttributeCount(), getPropertyCount() ));
+    UT_DEBUGMSG(("onlyContainsAbiwordChangeTrackingMarkup(top) ac:%ld pc:%ld\n",
+		 (long)getAttributeCount(), (long)getPropertyCount() ));
 
     if( !getAttributeCount() )
         return false;
